Implement Debug Manager and associated UI components
- Added DebugManager class to handle debug events and UI interactions.
- Created session-manager.ts to manage chat sessions and their states.
- Introduced tools-manager.ts for managing tools configuration.
- Developed main.ts to initialize the chat application and manage message sending.
- Updated index.html to include the new bundle.js script.
- Defined types for messages, tools, chat sessions, and debug events in types.ts.
- Adjusted tsconfig.json to support new TypeScript features and configurations.
