+++ src/ui/index.html
@@ -139,6 +139,6 @@
         </div>
     </div>
     
-    <script src="chat.js"></script>
+    <script src="bundle.js"></script>
 </body>
 </html>
